fix wav to wav logic

+++ convert.sh
@@ -13,7 +13,7 @@ if [[ $INPUT_FILE == *".ac3" ]]; then
 	ffmpeg -i $INPUT_FILE -vcodec copy -acodec pcm_s16le -ar $SAMPLING_FREQUENCY -ac $NUM_AUDIO_CHANNELS $DEST_FILE
 fi
 
-if [[ $INPUT_FILE == *".wav" ]]; then
+if [[ $INPUT_FILE == *".wav" ]] && [[ $DEST_FILE == *".wav" ]]; then
 	tmp_file=$INPUT_FILE.tmp.wav
 	ffmpeg -i $INPUT_FILE -ar $SAMPLING_FREQUENCY -ac $NUM_AUDIO_CHANNELS $tmp_file
 	rm $INPUT_FILE
